--- app/javascript/channels/index.js
@@ -1,5 +1,0 @@
-// Load all the channels within this directory and all subdirectories.
-// Channel files must be named *_channel.js.
-
-const channels = require.context('.', true, /_channel\.js$/);
-channels.keys().forEach(channels);
