+++ app/javascript/packs/hello_vue.js
@@ -9,6 +9,7 @@ import Vue from 'vue';
 import App from '../app.vue';
 import Vuetify from "vuetify";
 import "vuetify/dist/vuetify.min.css";
+import router from '../router'
 
 Vue.use(Vuetify);
 const vuetify = new Vuetify(
@@ -29,6 +30,7 @@ const vuetify = new Vuetify(
 document.addEventListener('DOMContentLoaded', () => {
   const app = new Vue({
     vuetify,
+    router,
     render: h => h(App)
   }).$mount();
   document.body.appendChild(app.$el);
