+++ app/javascript/router/index.js
@@ -1,0 +1,31 @@
+// Load all the channels within this directory and all subdirectories.
+// Channel files must be named *_channel.js.
+
+const channels = require.context('.', true, /_channel\.js$/);
+channels.keys().forEach(channels);
+import Vue from "vue";
+import Router from "vue-router";
+
+import TopIndex from "../pages/top/index";
+import SelectIndex from "../pages/select/index";
+
+Vue.use(Router)
+
+const router = new Router({
+  mode: "history",
+  routes: [
+    {
+      path: "/",
+      component: TopIndex,
+      name: "TopIndex",
+
+    },
+    {
+      path: "/select",
+      component: SelectIndex,
+      name: "SelectIndex",
+    },
+  ],
+})
+
+export default router
