+++ app/views/top/index.html.erb
@@ -1,8 +1,4 @@
-<<<<<<< HEAD
  <%= javascript_pack_tag 'hello_vue' %>
  <%= stylesheet_pack_tag 'hello_vue' %>
  
 <div id="app"></div>
-=======
- <%= javascript_pack_tag 'hello_vue' %>
->>>>>>> 9cfbdc7 ([Add]hello_vueの表示(#6))
