+++ package.json
@@ -9,6 +9,7 @@
     "turbolinks": "^5.2.0",
     "vue": "^2.6.14",
     "vue-loader": "^15.9.8",
+    "vue-router": "^3.5.3",
     "vue-template-compiler": "^2.6.14",
     "vuetify": "^2.6.2",
     "webpack": "^4.46.0",
